requirements/base.txt
-i https://pypi.org/simple
aiofiles==23.2.1; python_version >= '3.7'
aiogram==3.3.0; python_version >= '3.8'
aiohttp==3.9.3; python_version >= '3.8'
aiohttp-socks==0.8.4
aiosignal==1.3.1; python_version >= '3.7'
amqp==5.2.0; python_version >= '3.6'
aniso8601==9.0.1
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
billiard==4.2.0; python_version >= '3.7'
black==22.12.0; python_version >= '3.7'
celery==5.3.6; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
channels==4.0.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
contourpy==1.2.0; python_version >= '3.9'
cron-descriptor==1.4.3
cycler==0.12.1; python_version >= '3.8'
distro==1.9.0; python_version >= '3.6'
django==4.2.10; python_version >= '3.8'
django-axes[ipware]==6.3.0; python_version >= '3.7'
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==8.0.0
django-cors-headers==4.3.1; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-filter==23.5; python_version >= '3.7'
django-graphql-jwt==0.4.0; python_version >= '3.6'
django-redis==5.4.0; python_version >= '3.6'
django-timezone-field==6.1.0; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-spectacular==0.27.1; python_version >= '3.6'
exceptiongroup==1.2.0; python_version < '3.11'
flower==2.0.1; python_version >= '3.7'
fonttools==4.48.1; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.3; python_version >= '3.8'
httpx==0.26.0; python_version >= '3.8'
humanize==4.9.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jdatetime==3.8.2
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kiwisolver==1.4.5; python_version >= '3.7'
kombu==5.3.5; python_version >= '3.8'
lxml==4.9.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
magic-filter==1.0.12; python_version >= '3.7'
matplotlib==3.8.3; python_version >= '3.9'
mplfinance==0.12.10b0
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
openai==1.12.0; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
pandas==2.1.4; python_version >= '3.9'
pathspec==0.12.1; python_version >= '3.8'
pillow==10.1.0; python_version >= '3.8'
platformdirs==4.2.0; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
promise==2.3
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
pycparser==2.21
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-crontab==3.0.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-socks[asyncio]==2.4.4
pytse-client==0.18.0; python_full_version >= '3.7.2' and python_full_version < '4.0.0'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.1; python_version >= '3.7'
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.18.0; python_version >= '3.8'
scipy==1.12.0; python_version >= '3.9'
setuptools==69.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
ta-lib==0.4.28
tenacity==7.0.0
text-unidecode==1.3
tomli==2.0.1; python_full_version < '3.11.0a7'
tornado==6.4; python_version >= '3.8'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.0; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
whitenoise==6.6.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
